requirements.txt
firebase-admin==6.0.1
paho-mqtt==1.6.1

